Delete LICENSES/CC-BY-2.5.txt (#581)
As for the use of the CC-BY-2.5 license - it is only applicable to the gitflow.svg file, which was used in the contributor.md file. Since the contributor.md file is now part of the .github repo and no longer in the OpenSTEF repo, we can certainly remove the CC-BY-2.5 file from the LICENSES directory in OpenSTEF repo.

Signed-off-by: Jonas van den Bogaard <3628277+Jonasvdbo@users.noreply.github.com>
Co-authored-by: lschilders <lars.schilders@alliander.com>